% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_chains.R
\name{cat.chains}
\alias{cat.chains}
\title{A function to concatenate chains from two coda objects. The coda objects
must have chains with the same number of columns and the same length. The 
parameters in each chain must also match.}
\usage{
cat.chains(coda1, coda2)
}
\arguments{
\item{coda1}{A coda object form a model run.}

\item{coda2}{A coda object from a second model run.}
}
\description{
A function to concatenate chains from two coda objects. The coda objects
must have chains with the same number of columns and the same length. The 
parameters in each chain must also match.
}
