% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCrestart.R
\name{initfind}
\alias{initfind}
\title{A function to find initial values for a JAGS or OpenBUGS model.}
\usage{
initfind(coda, iteration = 0, OpenBUGS = FALSE)
}
\arguments{
\item{coda}{an mcmc object or list.}

\item{iteration}{the iteration to pull values from, defaults to the final iteration.}

\item{OpenBUGS}{Are the initials for OpenBUGS? Defaults to FALSE.

Output:
The output from this function is a list containing two elements. The first
contains the names of the variables and their indicies. These are useful 
when using removevars to remove variables that don't need initial values
in JAGS. The second element contains a list of initial values (this is a 
list of lists).
initfind()}
}
\description{
A function to find initial values for a JAGS or OpenBUGS model.
}
