% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_burnin.R
\name{mcmc.subset}
\alias{mcmc.subset}
\title{A function to remove burnin}
\usage{
mcmc.subset(coda, start = 1, end = 0)
}
\arguments{
\item{coda}{An mcmc object.}

\item{start}{The iteration to start on.}

\item{end}{The iteration to stop at defaults to the final iteration.
mcmc.subset()}
}
\description{
A function to remove burnin
}
