% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCrestart.R
\name{removevars}
\alias{removevars}
\title{A function to remove variables that don't need initial values in JAGS.}
\usage{
removevars(initsin, variables)
}
\arguments{
\item{initsin}{The initial values from the initsfind function.}

\item{variables}{The indexes of variables to be removed. Variables with 
indexes are available in the first list element from initsfind.
removevars()}
}
\description{
A function to remove variables that don't need initial values in JAGS.
}
