% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_chains.R
\name{stackchains}
\alias{stackchains}
\title{A function to stack chains in two coda objects. The objects must have the 
same number of chains each with the same number of columns.  The parameters 
in each chain must also match.}
\usage{
stackchains(coda1, coda2)
}
\arguments{
\item{coda1}{A coda object form a model run.}

\item{coda2}{A coda object form a second model run.}
}
\description{
A function to stack chains in two coda objects. The objects must have the 
same number of chains each with the same number of columns.  The parameters 
in each chain must also match.
}
