% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_fast.R
\name{coda.fast}
\alias{coda.fast}
\title{A function to summarize output from a JAGS or OpenBUGS model.}
\usage{
coda.fast(coda = NULL, thin = 1, FUN = NULL, colname = "optfun", ...)
}
\arguments{
\item{coda}{A coda object form a model run.}

\item{thin}{The thinning interval. Defaults to 1 (no thinning).}

\item{FUN}{An optional function to run on the coda object.}

\item{colname}{A name or list of names for the output of the optional 
function.}

\item{...}{Optional parameters to FUN.}
}
\description{
A function to summarize output from a JAGS or OpenBUGS model.
}
\keyword{coda}
