% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_fast.R
\name{coda.fast}
\alias{coda.fast}
\title{A function to summarize output from a JAGS or OpenBUGS model.}
\usage{
coda.fast(chains, thin = 1, coda = NULL, FUN = NULL,
  colname = "optfun", ...)
}
\arguments{
\item{chains}{The number of chains}

\item{thin}{The thinning interval. Defaults to 1 (no thinning).}

\item{coda}{A coda object form a model run.}

\item{FUN}{An optional function to run on the coda object.}

\item{colname}{A name or list of names for the output of the optional 
function.}

\item{...}{Optional parameters to FUN.

Output:
The output from this function is a list containing two elements. The first
contains the names of the variables and their indices. These are useful 
when using removevars to remove variables that don't need initial values
in JAGS. The second element contains a list of initial values (this is a 
list of lists).
coda.fast()}
}
\description{
A function to summarize output from a JAGS or OpenBUGS model.
}
